Move SAM1 core segmentation modules into micro_sam/v1 submodule
