+++ micro_sam/v1/__init__.py
@@ -1,0 +1,1 @@
+"""SAM1 (Segment Anything v1) functionality for micro_sam."""
