+++ test/test_multi_dimensional_segmentation.py
@@ -13,7 +13,7 @@ except ImportError:
 class TestMultiDimensionalSegmentation(unittest.TestCase):
 
     def test_merge_instance_segmentation_3d(self):
-        from micro_sam.multi_dimensional_segmentation import merge_instance_segmentation_3d
+        from micro_sam.v1.multi_dimensional_segmentation import merge_instance_segmentation_3d
 
         n_slices = 5
         data = np.stack(n_slices * binary_blobs(512))
@@ -37,7 +37,7 @@ class TestMultiDimensionalSegmentation(unittest.TestCase):
             self.assertTrue(np.array_equal(ids0, np.unique(merged_seg[z])))
 
     def test_merge_instance_segmentation_3d_with_closing(self):
-        from micro_sam.multi_dimensional_segmentation import merge_instance_segmentation_3d
+        from micro_sam.v1.multi_dimensional_segmentation import merge_instance_segmentation_3d
 
         n_slices = 5
         data = np.stack(n_slices * binary_blobs(512))
@@ -67,7 +67,7 @@ class TestMultiDimensionalSegmentation(unittest.TestCase):
 
     @unittest.skipIf(Trackastra is None, "Requires trackastra")
     def test_track_across_frames(self):
-        from micro_sam.multi_dimensional_segmentation import track_across_frames, get_napari_track_data
+        from micro_sam.v1.multi_dimensional_segmentation import track_across_frames, get_napari_track_data
 
         n_slices = 5
         data = binary_blobs(512).astype("uint8")
